+++ src/main/java/synapse/dementia/global/config/security/customFilter/CustomAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package synapse.dementia.global.config.security;
+package synapse.dementia.global.config.security.customFilter;
 
 import java.io.IOException;
